Shuffle files around to avoid NFS unlinking issues.
When unlinking an open file, NFS will instead rename the
file to .nfsXXXX until all file handles are closed.
This prevents ~/.fresh/build/bin from being removed
as the fresh script is running from this location.

The workaround is to instead move the old build directory
out of the way and then clean it up once the new build
directory is in place.

We move the old script out of the way into a directory which
is not going away before unlinking. This gives NFS somewhere
safe to store the temp file and allows us to clean up.

Closes #37.